+++ read_config.h
@@ -6,10 +6,10 @@
 config_setting_t *add_group (config_setting_t *, const char *);
 
 void read_float (config_setting_t *, const char *, double *);
-void read_int (config_setting_t *, const char *, int *);
+void read_int (config_setting_t *, const char *, long *);
 
 void read_float_array (config_setting_t *, const char *, double (*array)[], unsigned int);
-void read_int_array (config_setting_t *, const char *, int (*array)[], unsigned int);
+void read_int_array (config_setting_t *, const char *, long (*array)[], unsigned int);
 
 int read_config (const char *, const char *);
